moved cdi list devices to api
